Adds Basket class and passing test

+++ VendingMachineTechTest/Credit.cs
@@ -3,16 +3,16 @@ namespace VendingMachineTechTest
 {
     public class Credit
     {
-        public double total;
+        public decimal total;
 
         public Credit()
         {
             total = 0;
         }
 
-        public void Add(double coin)
+        public void Add(decimal coin)
         {
-            if (coin % 0.50 == 0){
+            if (coin % 0.5m == 0){
              total = +coin;   
             } else {
                 Console.WriteLine("Incorrect change added.");
